+++ libs/deal-ticket/src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './get-default-order';
 export * from './is-market-in-auction';
-export * from './validate-amount';
 export * from './validate-expiration';
 export * from './validate-market-state';
 export * from './validate-market-trading-mode';
